+++ src/clj/lambdacd/runners.clj
@@ -14,9 +14,7 @@
     (not (:retrigger-mock-for-build-number (:final-result step-finished-msg)))
     (not (:rerun-for-retrigger step-finished-msg))))
 
-(def ^{:deprecated "0.13.1"
-       :doc "Deprecated, should be private"}
-  should-trigger-next-build?
+(def ^:private should-trigger-next-build?
   (every-pred is-first-step? is-not-retriggered?))
 
 (defmacro while-not-stopped