+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This release contains no new functionality but only cleans up the codebase and r
 * Removed functions in  `lambdacd.core` (deprecated since 0.9.5)
 * Made helper-functions in `lambdacd.ui.ui-page` private (use was deprecated since 0.13.1)
 * Made helper-functions in `lambdacd.presentation.pipeline-structure` private (use was deprecated since 0.13.1)
+* Made helper-functions in `lambdacd.runners` private (use was deprecated since 0.13.1)
 
 ## 0.13.5