+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This release contains no new functionality but only cleans up the codebase and r
 * Made helper-functions in `lambdacd.ui.ui-page` private (use was deprecated since 0.13.1)
 * Made helper-functions in `lambdacd.presentation.pipeline-structure` private (use was deprecated since 0.13.1)
 * Made helper-functions in `lambdacd.runners` private (use was deprecated since 0.13.1)
+* Made helper-functions in `lambdacd.presentation.pipeline-structure` private (use was deprecated since 0.13.1)
 
 ## 0.13.5