+++ CHANGELOG.md
@@ -8,7 +8,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 **House-Keeping Release**. 
 
-This release contains no new functionality but only cleans up the codebase and removes deprecated interfaces.
+This release contains no new functionality but only cleans up the dependencies, codebase and removes deprecated interfaces. 
+
+These are **breaking changes** so if you are still using deprecated features or rely on some very specific behavior, take care when upgrading.
+
+### Changed
+
+* LambdaCD now depends on Clojure 1.8 by default. Clojure 1.7 is no longer officially supported but might still work.  
 
 ### Removed